First deposit successful

+++ Project-0/src/main/java/com/revature/models/Accounts/Account.java
@@ -1,8 +1,9 @@
 package com.revature.models.Accounts;
 
+
 public interface Account {
 
-    long   accountNumber = 0;
+    double accountNumber = 0;
     double balance       = 0;
